Add script for fetching routing information from RIPEstat
This commit adds scripts to query RIPEstat's API and fetch BGP routing
information that may be helpful in identifying instances of internet
shutdowns or outages in a given country.

Additionally, remove logging to a file in ~ and just log to stdout for
now. We had planned to do this for the final cescout release but we can
figure out where to log and the proper permissions with a Debian package
so we will sort this out when we get there.

Change-Id: I06601be387fe6be0dd554b2e36a96248ab67a8e6