Avoid manually formatting the datetime object for OONI's query
For the OONI metadb query, the date was formatted to %Y-%m-%d to strip
the time information; this was done as we only cared about measurements
in a 24-hour period. However, our own usage indicates that we need the
time information as well.

This commit changes that by running the query for the time period as
specified by the user using `since' and `until', so as to respect the
user choice in case they want to restrict the query to specific periods
of time, even in the same day.

Additionally, this fixes a bug which results from the `since' and
`until' being equal for a particular day: since the time information was
stripped, the query would become `test_start_time >= %Y-m%-%d AND
test_start_time <= %Y-m%-%d', effectively returning no result at all.

Change-Id: If87381369115af53de02a5dae05a9b0104b8b6aa